Add calculator operation functionality
User can now enter a number, operator, then a second number and the
calculator will display the correct answer upon user pressing the equals
button. Functions were added so that the variables for the numbers and
the operator are updated when user selects the corresponding buttons.
Users cannot enter more than one operator, enter an operator before
inputting a first number, enter a second number before the operator is
entered or enter an operator after the operation is complete.